src/Transportation & Safety/Traffic Management/Traffic Sign Recognition/requirements.txt
tensorflow==2.10.0
keras==2.10.0
pandas==1.5.0
pillow==9.2.0
scikit-learn==1.1.2
matplotlib==3.6.0
opencv-python==4.6.0.66  # For future video processing